header fix

+++ MicroFinal-RaceDisplay/include/pixel_display.hpp
@@ -18,7 +18,7 @@ void startupMessage();
 void displayHeat(char heatNum);
 void raceStart();
 void carFinish(char laneNum);
-void finalTimes(unsigned long lane1Millis, unsigned long lane2Millis);
+void finalTimes(unsigned int lane1Millis, unsigned int lane2Millis);
 void timeOut();
 
 #endif
